Basic authentication..?
1) index.js, before rendering <Gateway />, first makes a fetch request
to /auth
2) server.js checks whether user is logged in via passport, and returns
the logged in user object (stored in req.user for every request made to
the (8080?) server)
3) <Gateway /> is rendered with user=user
4) if user=null, <Login /> is rendered. Otherwise, go to <App />
5) The above is kept after browser refresh. I'm not sure how to log out,
but if you open a session in incognito you see that you're establishing
a new session and have to login again.

Login: (queries.js - findByUsername)
- check username is in db
- rn password is hardcoded as "wow"
- sets req.user to db row if login successful

TODO:
- make findById and change passport serialize/deserialize
- is it necessary for index.js to make a fetch request to /auth every
time (only on refresh???) Is this frequency enough??
	- where is this data stored?
	- how to log out?

- More generally, it seems all routing in terms of what page to render
with what props set are all done on CLIENT SIDE inside react-dom-router.
How does this interact with server side passport & express routing? Do I
always need to make a fetch request and then examine the data on client
side, since server-side redirecting seem to not work?

- Does fetch() work from client to server when deployed? It seems that
we need to give full IP address.

- Combine login page with the choose-room page in <App>

- make password actually stored in db