figured out why cors request & cookies were not being set properly. On the sender side, credentials: include is necessary to include cookies. On server side, Access-Control-Allow-Origin is necessary for browser to not block response, as well as enabling credentials. Both of these can be set in the `cors()` package, which then sets those headers on the HTTP response. These headers enable the browser to accept the response as valid, pass it to the code, and set the cookie in the session. `credentials` == cookies
All of this is b/c we're making a request from IP:3000 to IP:8080, and
by default the browser blocks cross-site (including diff ports) requests
, do not include cookies in the request, and do not allow cookies to be
set on the response.

Also, ENV var is being set on react-scripts start. The reason it's not being set in server.js is b/c that is run with pm2 // nodemon. To get it to work on production, I've updated pm2 start to first export the env var, like this:

`NODE_ENV=production pm2 restart server.js --update-env`